vhost-user: bump vhost/vm-memory/vmm-sys-util pins to windows-raw HEAD
vhost was 17 commits behind windows-raw (444a1c47 -> eff66ea), and the
vm-memory/vmm-sys-util [patch] pins were 15 commits behind the fork's
own pin for the same crates -- a stale [patch] next to a moved
git-branch dependency risks two incompatible compiled copies. Bumped
both in lockstep.

vhost's mmap_handle field changed from a pointer-typed handle to
RawDescriptor (isize on Windows) over that range, which broke the
Windows build; added the missing cast.

+++ vm/devices/virtio/vhost_user_frontend/src/lib.rs
@@ -483,7 +483,7 @@ impl VirtioDevice for VhostUserFrontend {
                         #[cfg(windows)]
                         let raw = {
                             use std::os::windows::io::AsRawHandle;
-                            region.file.as_raw_handle()
+                            region.file.as_raw_handle() as isize
                         };
                         VhostUserMemoryRegionInfo {
                             guest_phys_addr: region.guest_address,
